SERVER/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
passlib
python-jose
